requirements.txt
requests
requests-toolbelt
beautifulsoup4
pysimdjson
unitypackage_extractor
Pillow
pytz
boto3